split out download functionality
