+++ app.js
@@ -1,7 +1,8 @@
 import { caveLion, lightbearer, berserker, ironclad } from './src/exampleAssets.js'
-import { createAssetHtml } from './src/asset.js'
+import { createAssetHtml, setSvgDimensions } from './src/asset.js'
 import { showScreen } from './src/router.js'
 import { assetScale } from './src/assetScaling.js'
+import './src/download.js'
 
 const assetInput = document.querySelector('.interface-input')
 const assetContainer = document.querySelector(".assets")
@@ -9,9 +10,6 @@ const updateButton = document.querySelector(".update")
 const showLightbearerAssetButton = document.querySelector("#lightbearer-example")
 const showIroncladAssetButton = document.querySelector("#ironclad-example")
 const showCaveLionAssetButton = document.querySelector("#cave-lion-example")
-const downloadButton = document.querySelector("#download")
-const closeDownloadbutton = document.querySelector("#done-downloading")
-const downloadContainer = document.querySelector(".image-container")
 
 const showSingleAssetExample = (asset) => {
     assetInput.value = JSON.stringify(asset, null, 2)
@@ -23,52 +21,6 @@ const showHelpButton = document.querySelector('#show-help')
 showHelpButton.onclick = () => showScreen("help")
 const closeHelpbutton = document.querySelector("#close-help")
 closeHelpbutton.onclick = () => showScreen("main")
-
-
-const saveImage = (uri, filename) => {
-    const link = document.createElement('a')
-    link.href = uri
-    link.download = filename
-    document.body.appendChild(link)
-    link.click()
-    document.body.removeChild(link)
-}
-
-const setSvgDimensions = () => {
-    const svgs = document.querySelectorAll('svg')
-    svgs.forEach(svg => {
-        svg.setAttribute('height', svg.parentNode.offsetHeight)
-        svg.setAttribute('width', svg.parentNode.offsetWidth)
-    })
-}
-
-const screenshot = () => {
-    const render = document.querySelector('.render')
-    const asset = document.querySelector('.asset')
-    render.appendChild(asset)
-    showScreen('render')
-    window.scrollTo(0, 0)
-    setSvgDimensions()
-    html2canvas(asset,
-        {
-            allowTaint: true
-            // width: "750",
-            // height: "1050",
-            // scale: 2
-            // windowHeight: "1080px",
-            // windowWidth: "1920px"
-        }).then(canvas => {
-            downloadContainer.appendChild(canvas)
-            // saveImage(canvas.toDataURL(), 'asset.png') //TODO name after asset's name
-            showScreen('download')
-            document.querySelector('.assets').appendChild(asset)
-        })
-}
-
-const closeDownload = () => {
-    downloadContainer.innerHTML = ""
-    showScreen('main')
-}
 
 updateButton.onclick = () => {
     let data = JSON.parse(assetInput.value)
@@ -84,10 +36,5 @@ updateButton.onclick = () => {
 showLightbearerAssetButton.onclick = () => showSingleAssetExample(lightbearer)
 showIroncladAssetButton.onclick = () => showSingleAssetExample(ironclad)
 showCaveLionAssetButton.onclick = () => showSingleAssetExample(caveLion)
-downloadButton.onclick = screenshot
-closeDownloadbutton.onclick = closeDownload
-
-
-
 
 showSingleAssetExample(ironclad)
